screen objectifying
- make screens separate objects
- fix voltage screen selection jumping
- fix log interval on shutdown sometimes turning off
