+++ platformio.ini
@@ -18,6 +18,7 @@ framework = arduino
 lib_extra_dirs = libraries
 board_build.partitions = odroid.csv
 monitor_speed = 115200
+;monitor_speed = 921600
 upload_speed = 921600
 build_type = debug
 monitor_filters = esp32_exception_decoder
