+++ smartpower3/header.cpp
@@ -19,6 +19,8 @@ void Header::init(uint16_t x, uint16_t y)
 	this->x = x;
 	this->y = y;
 
+	tft->fillRect(0, 0, 480, 52, BG_COLOR);
+
 	input_v->fnd_init(3, 1, true, x + 35, y, FG_COLOR, BG_COLOR, FND_FONT_16x32, 100);
 	input_v->pushValue(0);
 	input_v->fnd_update(true);
@@ -56,7 +58,7 @@ void Header::onLogging(void)
 	update_logging_icon = true;
 }
 
-void Header::possibleLogging()
+void Header::possibleLogging(void)
 {
 	flag_logging = CAN_USE;
 	update_logging_icon = true;
@@ -217,7 +219,7 @@ void Header::pushPower(uint16_t volt, uint16_t ampere, uint16_t watt)
 	updated = true;
 }
 
-void Header::setDebug()
+void Header::setDebug(void)
 {
 	debug++;
 }
