+++ smartpower3/inputmanager.cpp
@@ -7,29 +7,25 @@ Button::Button(uint8_t pin) : PIN(pin)
 	pinMode(pin, INPUT_PULLUP);
 }
 
-void Button::isr_pol(void) {
-	if (pressed == 0) {
-		if (digitalRead(PIN) == 0) {
-			debounceTimer = millis();
-			pressed = 1;
-		}
-	} else if (digitalRead(PIN) == 1) {
-		if (pressed == 2)
-			pressed = 0;
+void Button::isr_pol(void)
+{
+	if (pressed == 0 && digitalRead(PIN) == 0) {
+		debounceTimer = millis();
+		pressed = 1;
+	} else if (digitalRead(PIN) == 1 && pressed == 2) {
+		pressed = 0;
 	}
 }
 
-uint8_t Button::checkPressed(void) {
-	if (pressed == 1) {
-		if (millis() - debounceTimer >= DEBOUNCE_TIME) {
-			if (digitalRead(PIN) == 1) {
-				pressed = 2;
-				return 1;
-			}
-			if (millis() - debounceTimer >= 3000) {
-				pressed = 2;
-				return 2;
-			}
+uint8_t Button::checkPressed(void)
+{
+	if (pressed == 1 && (millis() - debounceTimer >= DEBOUNCE_TIME)) {
+		if (millis() - debounceTimer >= LONG_PRESS_TIME) {
+			pressed = 2;
+			return 2;
+		} else if (digitalRead(PIN) == 1) {
+			pressed = 2;
+			return 1;
 		}
 	}
 	return 0;
@@ -48,6 +44,7 @@ void countEncoder(void *dial)
 {
 	struct dial_t *tmp = (struct dial_t *)dial;
 	int8_t cnt;
+
 	for (;;) {
 		cnt = encoder.getCount();
 		if (cnt > 9) {
